pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docops-mcp-assistant"
version = "0.1.0"
description = "Demo DocOps MCP Assistant with Gradio UI, agent, and MCP-like services"
authors = [{ name = "Your Name" }]
requires-python = ">=3.10"

dependencies = [
    "gradio>=4.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.12.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests", "mcp-servers"]
python_files = "test_*.py"